feat(ui): extract GridLoader into a shared component
GridLoader was defined inline inside the assistant message footer.
Moving it to components/grid-loader.tsx lets the VM booting state and
any other long-running placeholder reuse the same animation.
