+++ apps/mesh/src/web/components/chat/message/assistant.tsx
@@ -29,6 +29,7 @@ import {
 import { addUsage, emptyUsageStats } from "@decocms/mesh-sdk";
 import { useOptionalChatStream } from "../context.tsx";
 import { LiveTimer } from "../../live-timer.tsx";
+import { GridLoader } from "../../grid-loader.tsx";
 import { formatDuration } from "../../../lib/format-time.ts";
 
 type ThinkingStage = "planning" | "thinking";
@@ -85,55 +86,6 @@ function TypingIndicator() {
           {config.label}...
         </span>
       </span>
-    </div>
-  );
-}
-
-const GRID_CELLS = [
-  { delay: 0 },
-  { delay: 100 },
-  { delay: 200 },
-  { delay: 100 },
-  { delay: 200 },
-  { delay: 200 },
-  { delay: 300 },
-  { delay: 300 },
-  { delay: 400 },
-];
-
-function GridLoader() {
-  const [cellColors] = useState(() => {
-    const chart = `var(--chart-${Math.ceil(Math.random() * 5)})`;
-    return GRID_CELLS.map(() =>
-      Math.random() < 0.6
-        ? "color-mix(in srgb, var(--muted-foreground) 25%, transparent)"
-        : chart,
-    );
-  });
-  return (
-    <div
-      className="grid"
-      style={{
-        gridTemplateColumns: "repeat(3, 3px)",
-        gap: "1.5px",
-        width: "fit-content",
-      }}
-    >
-      {GRID_CELLS.map(({ delay }, i) => (
-        <div
-          key={i}
-          className="rounded-[1px]"
-          style={
-            {
-              width: 3,
-              height: 3,
-              "--cell-color": cellColors[i],
-              animation: "grid-ripple 1s ease infinite",
-              animationDelay: `${delay}ms`,
-            } as React.CSSProperties
-          }
-        />
-      ))}
     </div>
   );
 }
